Implemented Camera test
Added two files that showcase the workings of a camera module, Hardware Coding.doc and a picture.

Implemented a method on how to code Putty into simply making the Raspberry Pi Camera Module 3 to take a picture.

If applied, this commit will add a word document containing the coding, as well as the picture proof.