Updated Document and added Code for Microphone to Speaker
Updated Hardware Coding.docx that showcases how to connect the USB microphone to the USB stereo, along with adding two picture examples of the device.

Tested the code to check if the microphone was detected to the speaker (it resulted with low audio)

If applied, this commit will modify Hardware Coding.docx, add microphone_to_speaker.py, and the two pictures of the device.